fix: a few psy-rd correctness issues
chore: standardize psy-rd related costs to 64-bit values

fix: revert psy-rd contribution on loop filter strength

Looks like no matter what the psy-rd strength, the encoder ends up
effectively disabling loop filter for too many frames, causing too
much blockiness.

fix: more psy-rd correctness issues

+++ Source/Lib/Codec/product_coding_loop.c
@@ -992,30 +992,30 @@ static void fast_loop_core_light_pd0(ModeDecisionCandidateBuffer *cand_bf, Pictu
             uint8_t                *src_y  = input_pic->buffer_y + input_origin_index;
             *(cand_bf->fast_cost)          = fn_ptr->vf(pred_y, ref_pic->stride_y, src_y, input_pic->stride_y, &sse);
         } else {
-            *(cand_bf->fast_cost) = (uint32_t)(svt_spatial_full_distortion_kernel_facade(input_pic->buffer_y,
-                                                                                  input_origin_index,
-                                                                                  input_pic->stride_y << 1,
-                                                                                  ref_pic->buffer_y,
-                                                                                  ref_origin_index,
-                                                                                  ref_pic->stride_y << 1,
-                                                                                  ctx->blk_geom->bwidth,
-                                                                                  ctx->blk_geom->bheight >> 1,
-                                                                                  ctx->hbd_md,
-                                                                                  cand_bf->cand->block_mi.mode,
-                                                                                  cand_bf->cand->block_mi.interinter_comp.type,
-                                                                                  pcs->temporal_layer_index,
-                                                                                  pcs->scs->static_config.spy_rd))
+            *(cand_bf->fast_cost) = svt_spatial_full_distortion_kernel_facade(input_pic->buffer_y,
+                                                                              input_origin_index,
+                                                                              input_pic->stride_y << 1,
+                                                                              ref_pic->buffer_y,
+                                                                              ref_origin_index,
+                                                                              ref_pic->stride_y << 1,
+                                                                              ctx->blk_geom->bwidth,
+                                                                              ctx->blk_geom->bheight >> 1,
+                                                                              ctx->hbd_md,
+                                                                              cand_bf->cand->block_mi.mode,
+                                                                              cand_bf->cand->block_mi.interinter_comp.type,
+                                                                              pcs->temporal_layer_index,
+                                                                              pcs->scs->static_config.spy_rd)
                 << 1;
-            *(cand_bf->fast_cost) += (uint32_t)(get_svt_psy_full_dist(input_pic->buffer_y,
-                                                                      input_origin_index,
-                                                                      input_pic->stride_y << 1,
-                                                                      ref_pic->buffer_y,
-                                                                      ref_origin_index,
-                                                                      ref_pic->stride_y << 1,
-                                                                      ctx->blk_geom->bwidth,
-                                                                      ctx->blk_geom->bheight >> 1,
-                                                                      ctx->hbd_md,
-                                                                      pcs->ppcs->scs->static_config.psy_rd))
+            *(cand_bf->fast_cost) += get_svt_psy_full_dist(input_pic->buffer_y,
+                                                           input_origin_index,
+                                                           input_pic->stride_y << 1,
+                                                           ref_pic->buffer_y,
+                                                           ref_origin_index,
+                                                           ref_pic->stride_y << 1,
+                                                           ctx->blk_geom->bwidth,
+                                                           ctx->blk_geom->bheight >> 1,
+                                                           ctx->hbd_md,
+                                                           pcs->ppcs->scs->static_config.psy_rd)
                 << 1;
         }
     } else {
@@ -1046,7 +1046,7 @@ static void fast_loop_core_light_pd0(ModeDecisionCandidateBuffer *cand_bf, Pictu
 // Light PD1 fast loop core; assumes luma only, 8bit only, and that SSD is not used.
 static void fast_loop_core_light_pd1(ModeDecisionCandidateBuffer *cand_bf, PictureControlSet *pcs,
                                      ModeDecisionContext *ctx, EbPictureBufferDesc *input_pic, BlockLocation *loc) {
-    uint32_t       luma_fast_dist;
+    uint64_t       luma_fast_dist;
     const uint32_t fast_lambda = ctx->fast_lambda_md[EB_8_BIT_MD];
 
     ModeDecisionCandidate *cand = cand_bf->cand;
@@ -1141,7 +1141,7 @@ static void obmc_trans_face_off(ModeDecisionCandidateBuffer *cand_bf, PictureCon
             }
             uint64_t obmc_fast_luma_rate = cand_bf->fast_luma_rate + inter_mode_bits_num_obmc -
                 inter_mode_bits_num_translation;
-            uint32_t luma_fast_dist;
+            uint64_t luma_fast_dist;
             // Take a copy of the simple-translation results
             uint64_t simple_translation_cost                 = *(cand_bf->fast_cost);
             uint64_t simple_translation_fast_luma_rate       = cand_bf->fast_luma_rate;
@@ -1156,7 +1156,7 @@ static void obmc_trans_face_off(ModeDecisionCandidateBuffer *cand_bf, PictureCon
 
             // Distortion
             if (ctx->mds0_ctrls.mds0_dist_type == SSD) {
-                luma_fast_dist = (uint32_t)(svt_spatial_full_distortion_kernel_facade(
+                luma_fast_dist = svt_spatial_full_distortion_kernel_facade(
                     input_pic->buffer_y,
                     input_origin_index,
                     input_pic->stride_y,
@@ -1169,17 +1169,17 @@ static void obmc_trans_face_off(ModeDecisionCandidateBuffer *cand_bf, PictureCon
                     cand_bf->cand->block_mi.mode,
                     cand_bf->cand->block_mi.interinter_comp.type,
                     pcs->temporal_layer_index,
-                    pcs->scs->static_config.spy_rd));
-                luma_fast_dist += (uint32_t)(get_svt_psy_full_dist(input_pic->buffer_y,
-                                                                   input_origin_index,
-                                                                   input_pic->stride_y,
-                                                                   pred->buffer_y,
-                                                                   cu_origin_index,
-                                                                   pred->stride_y,
-                                                                   ctx->blk_geom->bwidth,
-                                                                   ctx->blk_geom->bheight,
-                                                                   ctx->hbd_md,
-                                                                   pcs->scs->static_config.psy_rd));
+                    pcs->scs->static_config.spy_rd);
+                luma_fast_dist += get_svt_psy_full_dist(input_pic->buffer_y,
+                                                        input_origin_index,
+                                                        input_pic->stride_y,
+                                                        pred->buffer_y,
+                                                        cu_origin_index,
+                                                        pred->stride_y,
+                                                        ctx->blk_geom->bwidth,
+                                                        ctx->blk_geom->bheight,
+                                                        ctx->hbd_md,
+                                                        pcs->scs->static_config.psy_rd);
                 cand_bf->luma_fast_dist = luma_fast_dist;
 
                 // Fast Cost
@@ -1236,7 +1236,7 @@ void fast_loop_core(ModeDecisionCandidateBuffer *cand_bf, PictureControlSet *pcs
                     EbPictureBufferDesc *input_pic, BlockLocation *loc) {
     const uint32_t input_origin_index = loc->input_origin_index;
     const uint32_t cu_origin_index    = loc->blk_origin_index;
-    uint32_t       luma_fast_dist;
+    uint64_t       luma_fast_dist;
     uint32_t       full_lambda  = ctx->hbd_md ? ctx->full_lambda_md[EB_10_BIT_MD] : ctx->full_lambda_md[EB_8_BIT_MD];
     uint32_t       fast_lambda  = ctx->hbd_md ? ctx->fast_lambda_md[EB_10_BIT_MD] : ctx->fast_lambda_md[EB_8_BIT_MD];
     uint32_t       lambda       = (ctx->mds0_ctrls.mds0_dist_type == SSD) ? full_lambda : fast_lambda;
@@ -1251,24 +1251,24 @@ void fast_loop_core(ModeDecisionCandidateBuffer *cand_bf, PictureControlSet *pcs
     if (ctx->mds0_ctrls.mds0_dist_type == SSD) {
         EbSpatialFullDistType spatial_full_dist_type_fun = ctx->hbd_md ? svt_full_distortion_kernel16_bits
                                                                        : svt_spatial_full_distortion_kernel;
-        luma_fast_dist = (uint32_t)(spatial_full_dist_type_fun(input_pic->buffer_y,
-                                                               input_origin_index,
-                                                               input_pic->stride_y,
-                                                               pred->buffer_y,
-                                                               (int32_t)cu_origin_index,
-                                                               pred->stride_y,
-                                                               ctx->blk_geom->bwidth,
-                                                               ctx->blk_geom->bheight));
-        luma_fast_dist += (uint32_t)(get_svt_psy_full_dist(input_pic->buffer_y,
-                                                           input_origin_index,
-                                                           input_pic->stride_y,
-                                                           pred->buffer_y,
-                                                           cu_origin_index,
-                                                           pred->stride_y,
-                                                           ctx->blk_geom->bwidth,
-                                                           ctx->blk_geom->bheight,
-                                                           ctx->hbd_md,
-                                                           pcs->scs->static_config.psy_rd));
+        luma_fast_dist = spatial_full_dist_type_fun(input_pic->buffer_y,
+                                                    input_origin_index,
+                                                    input_pic->stride_y,
+                                                    pred->buffer_y,
+                                                    (int32_t)cu_origin_index,
+                                                    pred->stride_y,
+                                                    ctx->blk_geom->bwidth,
+                                                    ctx->blk_geom->bheight);
+        luma_fast_dist += get_svt_psy_full_dist(input_pic->buffer_y,
+                                                input_origin_index,
+                                                input_pic->stride_y,
+                                                pred->buffer_y,
+                                                cu_origin_index,
+                                                pred->stride_y,
+                                                ctx->blk_geom->bwidth,
+                                                ctx->blk_geom->bheight,
+                                                ctx->hbd_md,
+                                                pcs->scs->static_config.psy_rd);
         cand_bf->luma_fast_dist = luma_fast_dist;
     } else {
         assert(ctx->mds0_ctrls.mds0_dist_type == VAR);
@@ -1874,7 +1874,7 @@ static void md_full_pel_search(PictureControlSet *pcs, ModeDecisionContext *ctx,
                                               : ctx->fast_lambda_md[hbd_md ? EB_10_BIT_MD : EB_8_BIT_MD];
     svt_init_mv_cost_params(
         &mv_cost_params, ctx, &ctx->ref_mv, frm_hdr->quantization_params.base_q_idx, rdmult, hbd_md);
-    uint32_t cost;
+    uint64_t cost;
     // Search area adjustment
     if ((ctx->blk_org_x + (mvx >> 3) + search_position_start_x) < (-ref_pic->org_x + 1))
         search_position_start_x = (-ref_pic->org_x + 1) - (ctx->blk_org_x + (mvx >> 3));
@@ -1951,24 +1951,24 @@ static void md_full_pel_search(PictureControlSet *pcs, ModeDecisionContext *ctx,
                 EbSpatialFullDistType spatial_full_dist_type_fun = hbd_md ? svt_full_distortion_kernel16_bits
                                                                           : svt_spatial_full_distortion_kernel;
 
-                cost = (uint32_t)spatial_full_dist_type_fun(input_pic->buffer_y,
-                                                            input_origin_index,
-                                                            input_pic->stride_y,
-                                                            ref_pic->buffer_y,
-                                                            ref_origin_index,
-                                                            ref_pic->stride_y,
-                                                            ctx->blk_geom->bwidth,
-                                                            ctx->blk_geom->bheight);
-                cost += (uint32_t)get_svt_psy_full_dist(input_pic->buffer_y,
-                                                         input_origin_index,
-                                                         input_pic->stride_y,
-                                                         ref_pic->buffer_y,
-                                                         ref_origin_index,
-                                                         ref_pic->stride_y,
-                                                         ctx->blk_geom->bwidth,
-                                                         ctx->blk_geom->bheight,
-                                                         ctx->hbd_md,
-                                                         pcs->scs->static_config.psy_rd);
+                cost = spatial_full_dist_type_fun(input_pic->buffer_y,
+                                                  input_origin_index,
+                                                  input_pic->stride_y,
+                                                  ref_pic->buffer_y,
+                                                  ref_origin_index,
+                                                  ref_pic->stride_y,
+                                                  ctx->blk_geom->bwidth,
+                                                  ctx->blk_geom->bheight);
+                cost += get_svt_psy_full_dist(input_pic->buffer_y,
+                                              input_origin_index,
+                                              input_pic->stride_y,
+                                              ref_pic->buffer_y,
+                                              ref_origin_index,
+                                              ref_pic->stride_y,
+                                              ctx->blk_geom->bwidth,
+                                              ctx->blk_geom->bheight,
+                                              ctx->hbd_md,
+                                              pcs->scs->static_config.psy_rd);
             } else {
                 assert((ctx->blk_geom->bwidth >> 3) < 17);
                 /* Might be able to add PSY distortion here too */
@@ -8424,7 +8424,7 @@ static void lpd1_tx_shortcut_detector(PictureControlSet *pcs, ModeDecisionContex
                                       ModeDecisionCandidateBuffer **cand_bf_ptr_array) {
     const BlockGeom             *blk_geom           = ctx->blk_geom;
     const ModeDecisionCandidate *cand               = cand_bf_ptr_array[ctx->mds0_best_idx]->cand;
-    const uint32_t               best_md_stage_dist = cand_bf_ptr_array[ctx->mds0_best_idx]->luma_fast_dist;
+    const uint64_t               best_md_stage_dist = cand_bf_ptr_array[ctx->mds0_best_idx]->luma_fast_dist;
     const uint32_t               th_normalizer      = blk_geom->bheight * blk_geom->bwidth * (pcs->picture_qp >> 1);
     ctx->use_tx_shortcuts_mds3                      = (100 * best_md_stage_dist) <
         (ctx->lpd1_tx_ctrls.use_mds3_shortcuts_th * th_normalizer);
@@ -8645,7 +8645,7 @@ static void md_encode_block_light_pd1(PictureControlSet *pcs, ModeDecisionContex
 void tx_shortcut_detector(PictureControlSet *pcs, ModeDecisionContext *ctx,
                           ModeDecisionCandidateBuffer **cand_bf_ptr_array) {
     const BlockGeom *blk_geom           = ctx->blk_geom;
-    const uint32_t   best_md_stage_dist = cand_bf_ptr_array[ctx->mds0_best_idx]->luma_fast_dist;
+    const uint64_t   best_md_stage_dist = cand_bf_ptr_array[ctx->mds0_best_idx]->luma_fast_dist;
     const uint32_t   th_normalizer      = blk_geom->bheight * blk_geom->bwidth * (pcs->picture_qp >> 1);
     ctx->use_tx_shortcuts_mds3          = (100 * best_md_stage_dist) <
         (ctx->tx_shortcut_ctrls.use_mds3_shortcuts_th * th_normalizer);
